+++ packages/react/src/components/Authenticator/ForceNewPassword/ForceNewPassword.tsx
@@ -8,7 +8,7 @@ import {
   getFormDataFromEvent,
 } from '../../../helpers/utils';
 import { useCustomComponents } from '../hooks/useCustomComponents';
-import { BaseFormFields } from '../shared/BaseFormFields';
+import { FormFields } from '../shared/FormFields';
 
 export const ForceNewPassword = (): JSX.Element => {
   const { error, isPending, toSignIn, submitForm, updateForm, updateBlur } =
@@ -91,4 +91,4 @@ export const ForceNewPassword = (): JSX.Element => {
   );
 };
 
-ForceNewPassword.FormFields = () => <BaseFormFields route="forceNewPassword" />;
+ForceNewPassword.FormFields = () => <FormFields route="forceNewPassword" />;
